point_2/stocks_products/requirements.txt
django~=4.1.2
djangorestframework~=3.14.0
django-filter

